Update: No devices in mount script

+++ Scripts/mount.sh
@@ -1,10 +1,22 @@
 #!/bin/bash
+# Get usb avilable block devices
 usbs=$(ls -lahF /sys/block | rg usb | cut -d '/' -f 12 | xargs -I {} lsblk --output NAME,SIZE,LABEL,MOUNTPOINT /dev/{} | rg 'sd[a-z]+[0-9]+' | sed 's/^..//')
 
-usb=$(echo "$usbs" | dmenu)
-usb_name=$(echo "$usb" | awk '{print $1}')
-usb_label=$(echo "$usb" | awk '{print $3}')
+# If There is no usb devices
+if [ -n "$usbs" ]; then
+	#Choose a usb device
+	usb=$(echo "$usbs" | dmenu -p "USB Devices: ")
+	# Get selected device name
+	usb_name=$(echo "$usb" | awk '{print $1}')
+	# Get Selected device label
+	usb_label=$(echo "$usb" | awk '{print $3}')
 
-[ -z "$usb" ] && exit
-[ ! -d "/mnt/USB/$usb_label" ] && doas mkdir "/mnt/USB/$usb_label" && doas chown adosha:adosha "/mnt/USB/$usb_label" -R && doas mount "/dev/$usb_name" "/mnt/USB/$usb_label" && exit
-[ -d "/mnt/USB/$usb_label" ] && doas umount "/dev/$usb_name" && doas rm -rf "/mnt/USB/$usb_label" && exit
+	# If no devices selected exit
+	[ -z "$usb" ] && exit
+	# Create a directory with selected device label and mount it to the directory
+	[ ! -d "/mnt/USB/$usb_label" ] && doas mkdir "/mnt/USB/$usb_label" && doas chown adosha:adosha "/mnt/USB/$usb_label" -R && doas mount "/dev/$usb_name" "/mnt/USB/$usb_label" && exit
+	# Unmount the selected device and remove the its directory
+	[ -d "/mnt/USB/$usb_label" ] && doas umount "/dev/$usb_name" && doas rm -rf "/mnt/USB/$usb_label" && exit
+else
+	dmenu -p "No USB Devices."
+fi
